feat: create a new ui activity and seperated old ui & new ui
