+++ app/src/main/kotlin/com/jeeldobariya/passcodes/oldui/AboutUsActivity.kt
@@ -1,4 +1,4 @@
-package com.jeeldobariya.passcodes.ui
+package com.jeeldobariya.passcodes.oldui
 
 import android.content.Intent
 import android.os.Bundle
