+++ app/src/main/kotlin/com/jeeldobariya/passcodes/oldui/LoadPasswordAction.kt
@@ -1,4 +1,4 @@
-package com.jeeldobariya.passcodes.ui
+package com.jeeldobariya.passcodes.oldui
 
 sealed interface LoadPasswordAction {
     data object RefreshPassword: LoadPasswordAction
