+++ app/src/main/kotlin/com/jeeldobariya/passcodes/oldui/SavePasswordActivity.kt
@@ -1,4 +1,4 @@
-package com.jeeldobariya.passcodes.ui
+package com.jeeldobariya.passcodes.oldui
 
 import android.os.Bundle
 import androidx.appcompat.app.AppCompatActivity
