+++ app/src/main/kotlin/com/jeeldobariya/passcodes/oldui/SavePasswordState.kt
@@ -1,10 +1,9 @@
-package com.jeeldobariya.passcodes.ui
+package com.jeeldobariya.passcodes.oldui
 
-data class ViewPasswordState(
+data class SavePasswordState(
     val domain: String = "",
     val username: String = "",
     val password: String = "",
     val notes: String = "",
-    val lastUpdatedAt: String = "",
     val isError: Boolean = false
 )
