+++ app/src/main/kotlin/com/jeeldobariya/passcodes/oldui/SettingsActivity.kt
@@ -1,6 +1,5 @@
-package com.jeeldobariya.passcodes.ui
+package com.jeeldobariya.passcodes.oldui
 
-import android.content.Context
 import android.os.Bundle
 import android.view.View
 import android.widget.AdapterView
