+++ app/src/main/kotlin/com/jeeldobariya/passcodes/oldui/UpdatePasswordActivity.kt
@@ -1,4 +1,4 @@
-package com.jeeldobariya.passcodes.ui
+package com.jeeldobariya.passcodes.oldui
 
 import android.app.AlertDialog
 import android.os.Bundle
