+++ app/src/main/kotlin/com/jeeldobariya/passcodes/oldui/ViewPasswordState.kt
@@ -1,9 +1,10 @@
-package com.jeeldobariya.passcodes.ui
+package com.jeeldobariya.passcodes.oldui
 
-data class SavePasswordState(
+data class ViewPasswordState(
     val domain: String = "",
     val username: String = "",
     val password: String = "",
     val notes: String = "",
+    val lastUpdatedAt: String = "",
     val isError: Boolean = false
 )
