+++ app/src/main/kotlin/com/jeeldobariya/passcodes/oldui/ViewPasswordViewModel.kt
@@ -1,4 +1,4 @@
-package com.jeeldobariya.passcodes.ui
+package com.jeeldobariya.passcodes.oldui
 
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
