+++ app/src/main/kotlin/com/jeeldobariya/passcodes/ui/MainActivity.kt
@@ -2,79 +2,84 @@ package com.jeeldobariya.passcodes.ui
 
 import android.content.Intent
 import android.os.Bundle
-import androidx.appcompat.app.AppCompatActivity
-import androidx.core.view.WindowCompat
-import androidx.lifecycle.lifecycleScope
-import com.jeeldobariya.passcodes.BuildConfig
-import com.jeeldobariya.passcodes.databinding.ActivityMainBinding
-import com.jeeldobariya.passcodes.utils.CommonUtils
-import com.jeeldobariya.passcodes.utils.UpdateChecker
-import kotlinx.coroutines.Dispatchers
-import kotlinx.coroutines.launch
+import androidx.activity.ComponentActivity
+import androidx.activity.compose.setContent
+import androidx.activity.enableEdgeToEdge
+import androidx.compose.foundation.layout.Arrangement
+import androidx.compose.foundation.layout.Column
+import androidx.compose.foundation.layout.Spacer
+import androidx.compose.foundation.layout.fillMaxSize
+import androidx.compose.foundation.layout.padding
+import androidx.compose.material3.Button
+import androidx.compose.material3.MaterialTheme
+import androidx.compose.material3.Surface
+import androidx.compose.material3.Text
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Alignment
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.tooling.preview.Preview
+import androidx.compose.ui.unit.dp
+import androidx.compose.ui.unit.sp
+import com.jeeldobariya.passcodes.ui.ui.theme.PasscodesTheme
 
-// import com.jeeldobariya.passcodes.utils.Permissions
-
-class MainActivity : AppCompatActivity() {
-
-    // private lateinit var permissionsHandle: Permissions
-
-    private lateinit var binding: ActivityMainBinding
-
+class MainActivity : ComponentActivity() {
     override fun onCreate(savedInstanceState: Bundle?) {
-        CommonUtils.updateCurrTheme(this)
         super.onCreate(savedInstanceState)
-        binding = ActivityMainBinding.inflate(layoutInflater)
-        setContentView(binding.root)
-
-        lifecycleScope.launch(Dispatchers.IO) {
-            UpdateChecker.checkVersion(this@MainActivity, BuildConfig.VERSION_NAME)
-        }
-
-        // Add event onclick listener
-        addOnClickListenerOnButton()
-
-        // Make window fullscreen
-        WindowCompat.setDecorFitsSystemWindows(window, false)
-
-        /* Comment the code as permission is not need to write into app data dir.
-            // Check and request permission when the app is first opened
-            permissionsHandle = Permissions(this)
-            if (!permissionsHandle.checkPermission()) permissionsHandle.requestPermission()
-        */
-    }
-
-    /* Comment the code as permission is not need to write into app data dir.
-        override fun onRequestPermissionsResult(
-            requestCode: Int, permissions: Array<String>, grantResults: IntArray
-        ) {
-            super.onRequestPermissionsResult(requestCode, permissions, grantResults)
-            if (requestCode == Permissions.PERMISSION_REQUEST_CODE) {
-                if (permissionsHandle.isPermissionGranted(grantResults)) {
-                    // Permission has been granted
-                    Toast.makeText(this, getString(R.string.permission_granted), Toast.LENGTH_LONG).show()
-                } else {
-                    // Permission not granted
-                    Toast.makeText(this, getString(R.string.permission_denied), Toast.LENGTH_LONG).show()
+        enableEdgeToEdge()
+        setContent {
+            PasscodesTheme {
+                MainScreen {
+                    val loginIntent = Intent(this@MainActivity, com.jeeldobariya.passcodes.oldui.MainActivity::class.java)
+                    startActivity(loginIntent)
                 }
             }
         }
-    */
+    }
+}
 
-    // Added all the onclick event listeners
-    private fun addOnClickListenerOnButton() {
-        binding.passwordManagerBtn.setOnClickListener {
-            val passwordManagerIntent = Intent(this, PasswordManagerActivity::class.java)
-            startActivity(passwordManagerIntent)
+@Composable
+fun MainScreen(navigateToOldUi: () -> Unit) {
+    Surface(
+        color = MaterialTheme.colorScheme.surface,
+        modifier = Modifier.fillMaxSize(),
+        tonalElevation = 5.dp
+    ) {
+        Column(
+            modifier = Modifier
+                .fillMaxSize()
+                .padding(24.dp),
+            verticalArrangement = Arrangement.Center,
+            horizontalAlignment = Alignment.CenterHorizontally
+        ) {
+            Text("Passcodes", fontSize = 24.sp)
+            Text("You are on New Jetpack UI")
+
+            Spacer(Modifier.padding(12.dp))
+
+            Button(
+                onClick = navigateToOldUi
+            ) {
+                Text("Back To Old UI", fontSize = 20.sp)
+            }
+            Button(
+                onClick = navigateToOldUi
+            ) {
+                Text("Continue New UI", fontSize = 20.sp)
+            }
+
+            Spacer(Modifier.padding(12.dp))
+
+            Text("Jetpack UI Is Under Development", fontSize = 6.sp)
         }
+    }
+}
 
-        binding.settingBtn.setOnClickListener {
-            val settingIntent = Intent(this, SettingsActivity::class.java)
-            startActivity(settingIntent)
-        }
-
-        binding.aboutUsBtn.setOnClickListener {
-            val aboutUsIntent = Intent(this, AboutUsActivity::class.java)
-            startActivity(aboutUsIntent)
+@Preview(showBackground = true)
+@Composable
+fun GreetingPreview() {
+    PasscodesTheme {
+        MainScreen {
+            Unit
         }
     }
 }
