+++ app/src/main/kotlin/com/jeeldobariya/passcodes/ui/ui/theme/Color.kt
@@ -1,4 +1,4 @@
-package com.jeeldobariya.passcodes.newui.ui.theme
+package com.jeeldobariya.passcodes.ui.ui.theme
 
 import androidx.compose.ui.graphics.Color
