+++ app/src/main/kotlin/com/jeeldobariya/passcodes/ui/ui/theme/Theme.kt
@@ -1,4 +1,4 @@
-package com.jeeldobariya.passcodes.newui.ui.theme
+package com.jeeldobariya.passcodes.ui.ui.theme
 
 import android.os.Build
 import androidx.compose.foundation.isSystemInDarkTheme
