+++ app/src/main/res/layout/activity_load_password.xml
@@ -5,7 +5,7 @@
 	android:layout_height="match_parent"
 	android:gravity="center"
     android:orientation="vertical"
-	tools:context=".ui.LoadPasswordActivity"
+	tools:context=".oldui.LoadPasswordActivity"
 	android:padding="4sp" >
 
 	<com.google.android.material.textview.MaterialTextView
