+++ app/src/main/res/layout/activity_main.xml
@@ -4,7 +4,7 @@
     android:layout_height="match_parent"
     android:gravity="center"
     android:orientation="vertical"
-    tools:context=".ui.MainActivity"
+    tools:context=".oldui.MainActivity"
     android:padding="4sp" >
 
     <ImageView
