+++ app/src/main/res/layout/activity_password_manager.xml
@@ -4,7 +4,7 @@
     android:layout_height="match_parent"
     android:gravity="center"
     android:orientation="vertical"
-    tools:context=".ui.PasswordManagerActivity"
+    tools:context=".oldui.PasswordManagerActivity"
     android:padding="4sp" >
     
     <com.google.android.material.textview.MaterialTextView
