+++ app/src/main/res/layout/activity_save_password.xml
@@ -6,7 +6,7 @@
 	android:layout_height="match_parent"
 	android:gravity="center"
   android:orientation="vertical"
-	tools:context=".ui.SavePasswordActivity"
+	tools:context=".oldui.SavePasswordActivity"
   android:padding="4sp" >
   
   <LinearLayout
