+++ app/src/main/res/layout/activity_settings.xml
@@ -7,7 +7,7 @@
     android:layout_height="match_parent"
     android:orientation="vertical"
     android:padding="16dp"
-    tools:context=".ui.SettingsActivity">
+    tools:context=".oldui.SettingsActivity">
 
     <!-- Settings Headline -->
     <com.google.android.material.textview.MaterialTextView
