+++ app/src/main/res/layout/activity_update_password.xml
@@ -5,7 +5,7 @@
 	android:layout_height="match_parent"
 	android:gravity="center"
   android:orientation="vertical"
-	tools:context=".ui.UpdatePasswordActivity"
+	tools:context=".oldui.UpdatePasswordActivity"
   android:padding="4sp" >
   
   <LinearLayout
