+++ app/src/main/res/layout/activity_view_password.xml
@@ -5,7 +5,7 @@
     android:layout_height="match_parent"
     android:gravity="center"
     android:orientation="vertical"
-    tools:context=".ui.ViewPasswordActivity"
+    tools:context=".oldui.ViewPasswordActivity"
     android:padding="4sp">
     
   <LinearLayout
